backend/requirements.txt
fastapi>=0.115.9
uvicorn>=0.34.0
neo4j>=5.28.1
langgraph>=0.5.0
langchain>0.3.25
langchain-openai>=0.2.14
langchain-community>=0.3.12
python-dotenv>=1.1.0
pydantic>=2.10.6
pydantic-settings>=2.6.1
pandas>=2.2.2
numpy>=1.26.4
scikit-learn>=1.4.2
python-multipart>=0.0.20
aiohttp>=3.12.2
python-jose[cryptography]>=3.5.0
passlib[bcrypt]==1.7.4
bcrypt>=4.3.0 